requirements.txt
rembg[cli]
pillow
watchdog
onnxruntime
flask
werkzeug
pathlib
gunicorn
python-dotenv
requests
python-json-logger